// src/components/providers/mascot/enums/method.enum.ts
export enum MascotMethod {
    GAME_LIST = "Game.List",
    BANK_GROUP_SET = "BankGroup.Set",
    PLAYER_SET = "Player.Set",
    SESSION_CREATE = "Session.Create",
    SESSION_CREATE_DEMO = "Session.CreateDemo",
    SESSION_CLOSE = "Session.Close",
    SESSION_GET = "Session.Get",
    SESSION_LIST = "Session.List",
    BONUS_SET = "Bonus.Set",
    BONUS_LIST = "Bonus.List",
    PLAYER_BONUS_LIST = "PlayerBonus.List",
    PLAYER_BONUS_GET = "PlayerBonus.Get",
    PLAYER_BONUS_ACTIVATE = "PlayerBonus.Activate",
    PLAYER_BONUS_DEACTIVATE = "PlayerBonus.Deactivate",
    PLAYER_BONUS_EXECUTE = "PlayerBonus.Execute",
    BALANCE_GET = "Balance.Get",
    BALANCE_CHANGE = "Balance.Change",
    SESSION_SPIN = "Session.Spin",
    JACKPOT_SET_SLOTS_VALUE = "Jackpot.SetSlotsValue"
}